curriculum-design-app/requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pandas==2.0.3
streamlit-option-menu==0.3.6
tenacity==8.2.3